Use hash-based service IDs
The hash is computed from service identifier (i.e. a combination
of service type, MAC address, port and path). Using hash-based
service IDs let us construct approximately the same service table
for a given set of services regardless of the detection order.
Note that the service table does not have to be always the same
for a fixed set of services due to possible hash collisions.