Create a file containing only Arrow client public identification (i.e. Arrow UUID) on application start
The file will be created only if requested by the corresponding
command line argument. It can be used in situations when the client
identification must be placed to an unsecure location (e.g. when it
needs to be accessed from a device web admin). Making the whole
configuration file accessible would disclose also the client's
secret key.