Fix service ID hashing instability caused by different pointer widths on 32-bit and 64-bit systems
This commit introduces a partially breaking change for 64-bit systems.
However, it shouldn't be a major issue. The breaking change would
only affect systems where the persistent service table is lost for
some reason. All services would get a different service IDs in such
case.

Given the fact that most of the deployments (if not all) are 32-bit
systems, this change should not cause any signifficant damage.